pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sensor-testing-control-server"
version = "0.1.0"
description = "control server for running the equipment in the QE closet for sensor testing"
readme = "README.md"
authors = []

dependencies = [
    "numpy",
    "scipy",
    "matplotlib",
    "pandas",
    "ipykernel",
    "labjack-ljm",
    "pyyaml",
    "pyserial",
    "pyro5",
    "pyvisa",
    "instrumentkit",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "black",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["labserver"]